[Setting] #94 - 미지원 버전 모두 제거
