+++ FFIP-iOS/FFIP-iOS/Sources/App/Factory/ModuleFactory.swift
@@ -9,21 +9,17 @@ import SwiftUI
 
 protocol ModuleFactoryProtocol {
     func makeExactCameraView(searchKeyword: String) -> ExactCameraView
-    @available(iOS 26.0, *)
-    func makeSemanticCameraView(searchKeyword: String) -> SemanticCameraView
     func makeSearchView(searchType: Binding<SearchType>) -> SearchView
-    @available(iOS 26.0, *)
     func makeVoiceSearchView(searchType: Binding<SearchType>) -> VoiceSearchView
-    func makeVoiceSearchSupportVersionVoew(searchType: Binding<SearchType>) -> VoiceSearchSupportVersionView
     func makePhotoDetailView() -> PhotoDetailView
     func makeOnboardingView() -> OnboardingView
 }
 
-final class ModuleFactory: ModuleFactoryProtocol {
-    static let shared = ModuleFactory()
+final class ModuleFactory: @preconcurrency ModuleFactoryProtocol {
+    @MainActor static let shared = ModuleFactory()
     private init() {}
     
-    func makeExactCameraView(searchKeyword: String) -> ExactCameraView {
+    @MainActor func makeExactCameraView(searchKeyword: String) -> ExactCameraView {
         let privacyService = PrivacyService()
         let captureService = VideoCaptureService()
         let deviceService = VideoDeviceService()
@@ -48,35 +44,8 @@ final class ModuleFactory: ModuleFactoryProtocol {
         )
         return view
     }
-    
-    @available(iOS 26.0, *)
-    func makeSemanticCameraView(searchKeyword: String) -> SemanticCameraView {
-        let privacyService = PrivacyService()
-        let captureService = VideoCaptureService()
-        let deviceService = VideoDeviceService()
-        let cameraModel = CameraModel(
-            privacyService: privacyService,
-            captureService: captureService,
-            deviceService: deviceService
-        )
-        
-        let visionService = VisionService()
-        let visionModel = VisionModel(searchKeyword: searchKeyword, visionService: visionService)
-        
-        let foundationModelsService = FoundationModelsService()
-        let languageModel = LanguageModel(foundationModelsService: foundationModelsService)
-        
-        let cameraMediator = SemanticCameraMediator(
-            cameraModel: cameraModel,
-            visionModel: visionModel,
-            languageModel: languageModel
-        )
-        
-        let view = SemanticCameraView(mediator: cameraMediator)
-        return view
-    }
-    
-    func makeSearchView(searchType: Binding<SearchType>) -> SearchView {
+
+    @MainActor func makeSearchView(searchType: Binding<SearchType>) -> SearchView {
         let keywords = UserDefaults.standard.stringArray(forKey: UserDefaultsKey.recentSearch) ?? []
         let model = SearchModel(recentSearchKeywords: keywords)
         let view = SearchView(
@@ -86,41 +55,21 @@ final class ModuleFactory: ModuleFactoryProtocol {
         return view
     }
     
-    @available(iOS 26.0, *)
-    func makeVoiceSearchView(searchType: Binding<SearchType>) -> VoiceSearchView {
-        let keywords = UserDefaults.standard.stringArray(forKey: UserDefaultsKey.recentSearch) ?? []
-        let searchModel = SearchModel(recentSearchKeywords: keywords)
-        
-        let privacyService = PrivacyService()
-        let speechTranscriptionService = SpeechTranscriptionService()
-        let voiceSearchModel = VoiceSearchModel(
-            privacyService: privacyService,
-            speechService: speechTranscriptionService
-        )
-        
-        let view = VoiceSearchView(
-            searchModel: searchModel,
-            voiceSearchModel: voiceSearchModel,
-            searchType: searchType
-        )
-        return view
-    }
-    
-    func makeVoiceSearchSupportVersionVoew(searchType: Binding<SearchType>) -> VoiceSearchSupportVersionView {
+    @MainActor func makeVoiceSearchView(searchType: Binding<SearchType>) -> VoiceSearchView {
         let privacyService = PrivacyService()
         let speechRecognitionService = SpeechRecognitionService()
-        let model = VoiceSearchSupportVersionModel(
+        let model = VoiceSearchModel(
             privacyService: privacyService,
             speechService: speechRecognitionService
         )
-        let view = VoiceSearchSupportVersionView(
+        let view = VoiceSearchView(
             voiceSearchSupportVersionModel: model,
             searchType: searchType
         )
         return view
     }
     
-    func makePhotoDetailView() -> PhotoDetailView {
+    @MainActor func makePhotoDetailView() -> PhotoDetailView {
         let view = PhotoDetailView()
         return view
     }
