+++ FFIP-iOS/FFIP-iOS/Sources/Presentation/Camera/VisionModel.swift
@@ -39,18 +39,6 @@ extension VisionModel {
     
     func processFrame(_ buffer: CVImageBuffer) async {
         do {
-            if #available(iOS 26.0, *) {
-                let isSmudge = try await visionService.performDetectSmudge(in: buffer, threshold: 0.95)
-                if isSmudge {
-                    countDetectSmudge += 1
-                    if countDetectSmudge > 5 {
-                        countDetectSmudge = 0
-                    }
-                    return
-                } else {
-                    countDetectSmudge = 0
-                }
-            }
             let textRects = try await visionService.performTextRecognition(image: buffer)
             self.recognizedTextObservations = textRects
         } catch {
@@ -60,11 +48,7 @@ extension VisionModel {
 
     func filterMatchedObservations() -> [RecognizedTextObservation] {
         recognizedTextObservations.filter {
-            if #available(iOS 26.0, *) {
-                $0.transcript.localizedCaseInsensitiveContains(searchKeyword)
-            } else {
-                $0.topCandidates(1).first?.string.localizedCaseInsensitiveContains(searchKeyword) ?? false
-            }
+            $0.topCandidates(1).first?.string.localizedCaseInsensitiveContains(searchKeyword) ?? false
         }
     }
 }
